Started a system builder class.
System builder is responsible for taking N files and forming it into one
concrete system.  It'll be used during the system construction phase so
that we can easily construct systems from numerous different types of
files.

Keep in mind that we'll ultimately have to handle collisions and such,
and allow previewing/scaling/moving/etc.  I wonder if there's a trivial
visualization method we could employ that could do this for us?

Something that would allow point/click modification of the starting
conditions, with a replay functionality.  We should probably center the
coordinates, as well; possibly on the COM of the molecule.

That'll be handled at the system builder phase, I suppose.  At least for
the initialization; the parser should center the molecule, if that's
what is necessary.